Split dom in different modules.
